+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Upcoming
 * Removed soul steel ingots from end city treasures
+* Part breaker is now a treasure enchantment
 
 ## 1.9.9
 * Spectral arrows are now made in the fletching table